some things are updated !

+++ src/app/modules/blog/blog.services.ts
@@ -69,11 +69,18 @@ const updateBlogFromDb = async (blogId : string , payload : TBlogs , token : Jwt
     return result ;
 }
 
-const deleteBlogFromDb = async (id : string) => {
+const deleteBlogFromDb = async (id : string , user : JwtPayload) => {
     const isBlogAxist = await blogsModel.findById(id) ;
+    const userData = await usersModel.findById(isBlogAxist?.author) ;
+    
     if(!isBlogAxist){
         throw new AppError(httpStatus.NOT_FOUND , "Blog not found !") ;
     }
+
+    if(user?.email !== userData?.email){
+        throw new AppError(httpStatus.UNAUTHORIZED , "You are not authorized !") ;
+    }
+
     const result = await blogsModel.findByIdAndDelete(id) ;
     return result ;
 }
